requirements.txt
# In general libraries should not be pinned to specific versions
attrs
h5py
git+https://github.com/ess-dmsc/python-nexus-utilities@6f5d2c7f1ca66063227aad2a111a2cc567d573f6#egg=nexusutils
git+https://github.com/ess-dmsc/nexus-json#egg=nexusjson
numpy-stl
pint
silx
xmltodict
numpy
pytest-qt

# PySide2 has had some issues with new versions and packaging so we will pin it
PySide2==5.13.0
